requirements.txt
kubernetes~=26.1.0
requests~=2.28.2
PyYAML~=6.0